@@ -4,7 +4,7 @@ version: 2.7.3
 homepage: https://github.com/PhilipsHue/flutter_reactive_ble
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0-133.2.beta <3.0.0'
   flutter: ">=1.10.0"
 
 dependencies:
