@@ -5,34 +5,34 @@ import 'package:meta/meta.dart';
 abstract class ConnectedDeviceOperation {
   Stream<CharacteristicValue> get characteristicValueStream;
 
-  Future<List<int>/*!*/> readCharacteristic(QualifiedCharacteristic/*!*/ characteristic);
+  Future<List<int>> readCharacteristic(QualifiedCharacteristic characteristic);
 
   Future<void> writeCharacteristicWithResponse(
-    QualifiedCharacteristic/*!*/ characteristic, {
-    @required List<int> value,
+    QualifiedCharacteristic characteristic, {
+    required List<int> value,
   });
 
   Future<void> writeCharacteristicWithoutResponse(
-    QualifiedCharacteristic/*!*/ characteristic, {
-    @required List<int> value,
+    QualifiedCharacteristic characteristic, {
+    required List<int> value,
   });
 
-  Stream<List<int>/*!*/> subscribeToCharacteristic(
-    QualifiedCharacteristic/*!*/ characteristic,
+  Stream<List<int>> subscribeToCharacteristic(
+    QualifiedCharacteristic characteristic,
     Future<void> isDisconnected,
   );
 
-  Future<int> requestMtu(String deviceId, int mtu);
+  Future<int> requestMtu(String deviceId, int? mtu);
 
   Future<List<DiscoveredService>> discoverServices(String deviceId);
 
   Future<void> requestConnectionPriority(
-      String deviceId, ConnectionPriority/*!*/ priority);
+      String deviceId, ConnectionPriority priority);
 }
 
 class ConnectedDeviceOperationImpl implements ConnectedDeviceOperation {
   ConnectedDeviceOperationImpl({
-    @required DeviceOperationController controller,
+    required DeviceOperationController controller,
   })  : assert(controller != null),
         _controller = controller;
 
@@ -43,22 +43,22 @@ class ConnectedDeviceOperationImpl implements ConnectedDeviceOperation {
       _controller.charValueUpdateStream;
 
   @override
-  Future<List<int>/*!*/> readCharacteristic(QualifiedCharacteristic characteristic) {
+  Future<List<int>> readCharacteristic(QualifiedCharacteristic characteristic) {
     final specificCharacteristicValueStream = characteristicValueStream
         .where((update) => update.characteristic == characteristic)
         .map((update) => update.result.dematerialize());
 
     return _controller
         .readCharacteristic(characteristic)
-        .asyncExpand((Object _) => specificCharacteristicValueStream)
+        .asyncExpand((Object? _) => specificCharacteristicValueStream)
         .firstWhere((_) => true,
-            orElse: () => throw NoBleCharacteristicDataReceived());
+            orElse: () => throw NoBleCharacteristicDataReceived()) as Future<List<int>>;
   }
 
   @override
   Future<void> writeCharacteristicWithResponse(
     QualifiedCharacteristic characteristic, {
-    @required List<int> value,
+    required List<int> value,
   }) async =>
       _controller
           .writeCharacteristicWithResponse(characteristic, value)
@@ -67,14 +67,14 @@ class ConnectedDeviceOperationImpl implements ConnectedDeviceOperation {
   @override
   Future<void> writeCharacteristicWithoutResponse(
     QualifiedCharacteristic characteristic, {
-    @required List<int> value,
+    required List<int> value,
   }) async =>
       _controller
           .writeCharacteristicWithoutResponse(characteristic, value)
           .then((info) => info.result.dematerialize());
 
   @override
-  Stream<List<int>/*!*/> subscribeToCharacteristic(
+  Stream<List<int>> subscribeToCharacteristic(
     QualifiedCharacteristic characteristic,
     Future<void> isDisconnected,
   ) {
@@ -82,10 +82,10 @@ class ConnectedDeviceOperationImpl implements ConnectedDeviceOperation {
         .where((update) => update.characteristic == characteristic)
         .map((update) => update.result.dematerialize());
 
-    final autosubscribingRepeater = Repeater<List<int>>.broadcast(
+    final autosubscribingRepeater = Repeater<List<int>?>.broadcast(
       onListenEmitFrom: () => _controller
           .subscribeToNotifications(characteristic)
-          .asyncExpand((Object _) => specificCharacteristicValueStream),
+          .asyncExpand((Object? _) => specificCharacteristicValueStream),
       onCancel: () => _controller
           .stopSubscribingToNotifications(characteristic)
           .catchError((Object e) =>
@@ -94,11 +94,11 @@ class ConnectedDeviceOperationImpl implements ConnectedDeviceOperation {
 
     isDisconnected.then<void>((_) => autosubscribingRepeater.dispose());
 
-    return autosubscribingRepeater.stream;
+    return autosubscribingRepeater.stream as Stream<List<int>>;
   }
 
   @override
-  Future<int> requestMtu(String deviceId, int mtu) async =>
+  Future<int> requestMtu(String deviceId, int? mtu) async =>
       _controller.requestMtuSize(deviceId, mtu);
 
   @override
@@ -110,7 +110,7 @@ class ConnectedDeviceOperationImpl implements ConnectedDeviceOperation {
           String deviceId, ConnectionPriority priority) async =>
       _controller
           .requestConnectionPriority(deviceId, priority)
-          .then((message) => message.result.dematerialize());
+          .then((message) => message.result!.dematerialize());
 }
 
 @visibleForTesting
