@@ -6,43 +6,43 @@ import 'package:meta/meta.dart';
 abstract class ArgsToProtobufConverter {
   pb.ConnectToDeviceRequest createConnectToDeviceArgs(
     String id,
-    Map<Uuid, List<Uuid>> servicesWithCharacteristicsToDiscover,
-    Duration connectionTimeout,
+    Map<Uuid, List<Uuid>>? servicesWithCharacteristicsToDiscover,
+    Duration? connectionTimeout,
   );
 
   pb.DisconnectFromDeviceRequest createDisconnectDeviceArgs(String deviceId);
 
   pb.ReadCharacteristicRequest createReadCharacteristicRequest(
-    QualifiedCharacteristic/*!*/ characteristic,
+    QualifiedCharacteristic characteristic,
   );
 
   pb.WriteCharacteristicRequest createWriteChacracteristicRequest(
-    QualifiedCharacteristic/*!*/ characteristic,
+    QualifiedCharacteristic characteristic,
     List<int> value,
   );
 
   pb.NotifyCharacteristicRequest createNotifyCharacteristicRequest(
-    QualifiedCharacteristic/*!*/ characteristic,
+    QualifiedCharacteristic characteristic,
   );
 
   pb.NotifyNoMoreCharacteristicRequest createNotifyNoMoreCharacteristicRequest(
-    QualifiedCharacteristic/*!*/ characteristic,
+    QualifiedCharacteristic characteristic,
   );
 
   pb.NegotiateMtuRequest createNegotiateMtuRequest(
     String deviceId,
-    int/*!*/ mtu,
+    int mtu,
   );
 
   pb.ChangeConnectionPriorityRequest createChangeConnectionPrioRequest(
     String deviceId,
-    ConnectionPriority/*!*/ priority,
+    ConnectionPriority priority,
   );
 
   pb.ScanForDevicesRequest createScanForDevicesRequest({
-    @required List<Uuid> withServices,
-    @required ScanMode scanMode,
-    @required bool requireLocationServicesEnabled,
+    required List<Uuid> withServices,
+    required ScanMode? scanMode,
+    required bool? requireLocationServicesEnabled,
   });
 
   pb.ClearGattCacheRequest createClearGattCacheRequest(String deviceId);
@@ -56,8 +56,8 @@ class ArgsToProtobufConverterImpl implements ArgsToProtobufConverter {
   @override
   pb.ConnectToDeviceRequest createConnectToDeviceArgs(
     String id,
-    Map<Uuid, List<Uuid>> servicesWithCharacteristicsToDiscover,
-    Duration connectionTimeout,
+    Map<Uuid, List<Uuid>>? servicesWithCharacteristicsToDiscover,
+    Duration? connectionTimeout,
   ) {
     final args = pb.ConnectToDeviceRequest()..deviceId = id;
 
@@ -69,7 +69,7 @@ class ArgsToProtobufConverterImpl implements ArgsToProtobufConverter {
       final items = <pb.ServiceWithCharacteristics>[];
       for (final serviceId in servicesWithCharacteristicsToDiscover.keys) {
         final characteristicIds =
-            servicesWithCharacteristicsToDiscover[serviceId];
+            servicesWithCharacteristicsToDiscover[serviceId]!;
         items.add(
           pb.ServiceWithCharacteristics()
             ..serviceId = (pb.Uuid()..data = serviceId.data)
@@ -89,7 +89,7 @@ class ArgsToProtobufConverterImpl implements ArgsToProtobufConverter {
 
   @override
   pb.ReadCharacteristicRequest createReadCharacteristicRequest(
-    QualifiedCharacteristic/*!*/ characteristic,
+    QualifiedCharacteristic characteristic,
   ) {
     final args = pb.ReadCharacteristicRequest()
       ..characteristic = (pb.CharacteristicAddress()
@@ -171,13 +171,13 @@ class ArgsToProtobufConverterImpl implements ArgsToProtobufConverter {
 
   @override
   pb.ScanForDevicesRequest createScanForDevicesRequest({
-    @required List<Uuid> withServices,
-    @required ScanMode scanMode,
-    @required bool requireLocationServicesEnabled,
+    required List<Uuid>? withServices,
+    required ScanMode? scanMode,
+    required bool? requireLocationServicesEnabled,
   }) {
     final args = pb.ScanForDevicesRequest()
-      ..scanMode = convertScanModeToArgs(scanMode)
-      ..requireLocationServicesEnabled = requireLocationServicesEnabled;
+      ..scanMode = convertScanModeToArgs(scanMode!)
+      ..requireLocationServicesEnabled = requireLocationServicesEnabled!;
 
     if (withServices != null) {
       for (final withService in withServices) {
