@@ -13,29 +13,29 @@ import 'model/uuid.dart';
 abstract class DeviceConnector {
   Stream<ConnectionStateUpdate> get deviceConnectionStateUpdateStream;
 
-  Stream<ConnectionStateUpdate/*!*/> connect({
-    @required String id,
-    Map<Uuid, List<Uuid>> servicesWithCharacteristicsToDiscover,
-    Duration connectionTimeout,
+  Stream<ConnectionStateUpdate> connect({
+    required String id,
+    Map<Uuid, List<Uuid>>? servicesWithCharacteristicsToDiscover,
+    Duration? connectionTimeout,
   });
 
-  Stream<ConnectionStateUpdate/*!*/> connectToAdvertisingDevice({
-    @required String id,
-    @required List<Uuid> withServices,
-    @required Duration prescanDuration,
-    Map<Uuid, List<Uuid>> servicesWithCharacteristicsToDiscover,
-    Duration connectionTimeout,
+  Stream<ConnectionStateUpdate> connectToAdvertisingDevice({
+    required String id,
+    required List<Uuid> withServices,
+    required Duration prescanDuration,
+    Map<Uuid, List<Uuid>>? servicesWithCharacteristicsToDiscover,
+    Duration? connectionTimeout,
   });
 }
 
 class DeviceConnectorImpl implements DeviceConnector {
   const DeviceConnectorImpl({
-    @required DeviceConnectionController controller,
-    @required
-        bool Function({String deviceId, Duration cacheValidity})
+    required DeviceConnectionController controller,
+    required
+        bool Function({String? deviceId, Duration? cacheValidity})
             deviceIsDiscoveredRecently,
-    @required DeviceScanner deviceScanner,
-    @required Duration delayAfterScanFailure,
+    required DeviceScanner deviceScanner,
+    required Duration delayAfterScanFailure,
   })  : assert(controller != null),
         assert(deviceScanner != null),
         assert(deviceIsDiscoveredRecently != null),
@@ -46,7 +46,7 @@ class DeviceConnectorImpl implements DeviceConnector {
         _delayAfterScanFailure = delayAfterScanFailure;
 
   final DeviceConnectionController _controller;
-  final bool Function({String deviceId, Duration cacheValidity})
+  final bool Function({String? deviceId, Duration? cacheValidity})
       _deviceIsDiscoveredRecently;
   final DeviceScanner _deviceScanner;
   final Duration _delayAfterScanFailure;
@@ -58,10 +58,10 @@ class DeviceConnectorImpl implements DeviceConnector {
       _controller.connectionUpdateStream;
 
   @override
-  Stream<ConnectionStateUpdate> connect({
-    @required String id,
-    Map<Uuid, List<Uuid>> servicesWithCharacteristicsToDiscover,
-    Duration connectionTimeout,
+  Stream<ConnectionStateUpdate?> connect({
+    required String id,
+    Map<Uuid, List<Uuid>>? servicesWithCharacteristicsToDiscover,
+    Duration? connectionTimeout,
   }) {
     final specificConnectedDeviceStream = deviceConnectionStateUpdateStream
         .where((update) => update.deviceId == id)
@@ -75,7 +75,7 @@ class DeviceConnectorImpl implements DeviceConnector {
       onListenEmitFrom: () => _controller
           .connectToDevice(
               id, servicesWithCharacteristicsToDiscover, connectionTimeout)
-          .asyncExpand((Object _) => specificConnectedDeviceStream),
+          .asyncExpand((Object? _) => specificConnectedDeviceStream),
       onCancel: () => _controller.disconnectDevice(id),
     );
 
@@ -83,12 +83,12 @@ class DeviceConnectorImpl implements DeviceConnector {
   }
 
   @override
-  Stream<ConnectionStateUpdate> connectToAdvertisingDevice({
-    @required String id,
-    @required List<Uuid> withServices,
-    @required Duration prescanDuration,
-    Map<Uuid, List<Uuid>> servicesWithCharacteristicsToDiscover,
-    Duration connectionTimeout,
+  Stream<ConnectionStateUpdate?> connectToAdvertisingDevice({
+    required String id,
+    required List<Uuid> withServices,
+    required Duration prescanDuration,
+    Map<Uuid, List<Uuid>>? servicesWithCharacteristicsToDiscover,
+    Duration? connectionTimeout,
   }) {
     if (_deviceScanner.currentScan != null) {
       return _awaitCurrentScanAndConnect(
@@ -109,10 +109,10 @@ class DeviceConnectorImpl implements DeviceConnector {
     }
   }
 
-  Stream<ConnectionStateUpdate> _prescanAndConnect(
+  Stream<ConnectionStateUpdate?> _prescanAndConnect(
     String id,
-    Map<Uuid, List<Uuid>> servicesWithCharacteristicsToDiscover,
-    Duration connectionTimeout,
+    Map<Uuid, List<Uuid>>? servicesWithCharacteristicsToDiscover,
+    Duration? connectionTimeout,
     List<Uuid> withServices,
     Duration prescanDuration,
   ) {
@@ -133,7 +133,7 @@ class DeviceConnectorImpl implements DeviceConnector {
         scanSubscription.cancel();
       });
 
-      return _deviceScanner.currentScan.future
+      return _deviceScanner.currentScan!.future
           .then((_) => true)
           .catchError((Object _) => false)
           .asStream()
@@ -160,16 +160,16 @@ class DeviceConnectorImpl implements DeviceConnector {
     }
   }
 
-  Stream<ConnectionStateUpdate> _awaitCurrentScanAndConnect(
+  Stream<ConnectionStateUpdate?> _awaitCurrentScanAndConnect(
     List<Uuid> withServices,
     Duration prescanDuration,
     String id,
-    Map<Uuid, List<Uuid>> servicesWithCharacteristicsToDiscover,
-    Duration connectionTimeout,
+    Map<Uuid, List<Uuid>>? servicesWithCharacteristicsToDiscover,
+    Duration? connectionTimeout,
   ) {
     if (const DeepCollectionEquality()
-        .equals(_deviceScanner.currentScan.withServices, withServices)) {
-      return _deviceScanner.currentScan.future
+        .equals(_deviceScanner.currentScan!.withServices, withServices)) {
+      return _deviceScanner.currentScan!.future
           .timeout(prescanDuration + const Duration(seconds: 1))
           .asStream()
           .asyncExpand(
@@ -195,10 +195,10 @@ class DeviceConnectorImpl implements DeviceConnector {
     }
   }
 
-  Stream<ConnectionStateUpdate> _connectIfRecentlyDiscovered(
+  Stream<ConnectionStateUpdate?> _connectIfRecentlyDiscovered(
     String id,
-    Map<Uuid, List<Uuid>> servicesWithCharacteristicsToDiscover,
-    Duration connectionTimeout,
+    Map<Uuid, List<Uuid>>? servicesWithCharacteristicsToDiscover,
+    Duration? connectionTimeout,
   ) {
     if (_deviceIsDiscoveredRecently(
       deviceId: id,
