@@ -1,7 +1,7 @@
 import 'package:meta/meta.dart';
 
 abstract class DiscoveredDevicesRegistry {
-  DiscoveredDevicesRegistry({@required this.getTimestamp});
+  DiscoveredDevicesRegistry({required this.getTimestamp});
 
   final DateTime Function() getTimestamp;
 
@@ -11,11 +11,11 @@ abstract class DiscoveredDevicesRegistry {
 
   bool isEmpty();
 
-  bool deviceIsDiscoveredRecently({String deviceId, Duration cacheValidity});
+  bool deviceIsDiscoveredRecently({String? deviceId, Duration? cacheValidity});
 }
 
 class DiscoveredDevicesRegistryImpl implements DiscoveredDevicesRegistry {
-  DiscoveredDevicesRegistryImpl({@required this.getTimestamp});
+  DiscoveredDevicesRegistryImpl({required this.getTimestamp});
 
   DiscoveredDevicesRegistryImpl.standard()
       : this(getTimestamp: () => DateTime.now());
@@ -40,8 +40,8 @@ class DiscoveredDevicesRegistryImpl implements DiscoveredDevicesRegistry {
   bool isEmpty() => discoveredDevices.isEmpty;
 
   @override
-  bool deviceIsDiscoveredRecently({String deviceId, Duration cacheValidity}) =>
+  bool deviceIsDiscoveredRecently({String? deviceId, Duration? cacheValidity}) =>
       discoveredDevices.containsKey(deviceId) &&
-      discoveredDevices[deviceId]
-          .isAfter(getTimestamp().subtract(cacheValidity));
+      discoveredDevices[deviceId!]!
+          .isAfter(getTimestamp().subtract(cacheValidity!));
 }
