@@ -5,12 +5,12 @@ import 'package:meta/meta.dart';
 
 @immutable
 class WriteCharacteristicInfo {
-  final QualifiedCharacteristic/*!*/ characteristic;
-  final Result<void, GenericFailure<WriteCharacteristicFailure>/*!*/> result;
+  final QualifiedCharacteristic characteristic;
+  final Result<void, GenericFailure<WriteCharacteristicFailure>> result;
 
   const WriteCharacteristicInfo({
-    @required this.characteristic,
-    @required this.result,
+    required this.characteristic,
+    required this.result,
   });
 }
 
