@@ -1,7 +1,7 @@
 import 'package:meta/meta.dart';
 
 T selectFrom<T>(List<T> values,
-    {@required int index, @required T Function(int index) fallback}) {
+    {required int? index, required T Function(int? index) fallback}) {
   assert(values != null);
   assert(fallback != null);
 
