@@ -14,16 +14,16 @@ import 'device_connector_test.mocks.dart';
     [DeviceConnectionController, DeviceScanner, DiscoveredDevicesRegistry])
 void main() {
   group('$DeviceConnector', () {
-    DeviceConnector _sut;
-    MockDeviceConnectionController _controller;
-    Stream<ConnectionStateUpdate> _connectionStateUpdateStream;
-    Stream<ConnectionStateUpdate> _result;
-    MockDiscoveredDevicesRegistry _registry;
-    MockDeviceScanner _scanner;
+    late DeviceConnector _sut;
+    late MockDeviceConnectionController _controller;
+    Stream<ConnectionStateUpdate?>? _connectionStateUpdateStream;
+    Stream<ConnectionStateUpdate>? _result;
+    late MockDiscoveredDevicesRegistry _registry;
+    late MockDeviceScanner _scanner;
 
-    Map<Uuid, List<Uuid>> _servicesToDiscover;
-    ConnectionStateUpdate updateForDevice;
-    ConnectionStateUpdate updateOtherDevice;
+    Map<Uuid, List<Uuid>>? _servicesToDiscover;
+    ConnectionStateUpdate? updateForDevice;
+    ConnectionStateUpdate? updateOtherDevice;
 
     const _deviceId = '123';
     const _connectionTimeout = Duration(seconds: 1);
@@ -37,9 +37,9 @@ void main() {
         Uuid.parse('FEFE'): [Uuid.parse('FEFE')]
       };
       when(_controller.connectionUpdateStream)
-          .thenAnswer((_) => _connectionStateUpdateStream);
+          .thenAnswer(((_) => _connectionStateUpdateStream as Stream<ConnectionStateUpdate>) as Stream<ConnectionStateUpdate> Function(Invocation));
 
-      when(_controller.disconnectDevice(any)).thenAnswer((_) async => 0);
+      when(_controller.disconnectDevice(any!)).thenAnswer((_) async => 0);
 
       updateForDevice = const ConnectionStateUpdate(
         deviceId: _deviceId,
@@ -71,7 +71,7 @@ void main() {
 
         group('And invoking connect method succeeds', () {
           setUp(() async {
-            when(_controller.connectToDevice(any, any, any)).thenAnswer(
+            when(_controller.connectToDevice(any!, any, any)).thenAnswer(
               (_) => Stream.fromIterable([1]),
             );
             _result = _sut.connect(
@@ -83,7 +83,7 @@ void main() {
 
           test('It emits connection updates for that device', () {
             expect(_result,
-                emitsInOrder(<ConnectionStateUpdate>[updateForDevice]));
+                emitsInOrder(<ConnectionStateUpdate?>[updateForDevice]));
           });
         });
       });
@@ -155,7 +155,7 @@ void main() {
                     deviceId: deviceId,
                     cacheValidity: anyNamed('cacheValidity')))
                 .thenReturn(true);
-            when(_controller.connectToDevice(any, any, any))
+            when(_controller.connectToDevice(any!, any, any))
                 .thenAnswer((_) => Stream.fromIterable([1]));
 
             _result = _sut.connectToAdvertisingDevice(
@@ -167,7 +167,7 @@ void main() {
 
           test('It connects to device after scan has finished', () {
             expect(_result,
-                emitsInOrder(<ConnectionStateUpdate>[updateForDevice]));
+                emitsInOrder(<ConnectionStateUpdate?>[updateForDevice]));
           });
         });
 
@@ -216,7 +216,7 @@ void main() {
                     deviceId: deviceId,
                     cacheValidity: anyNamed('cacheValidity')))
                 .thenReturn(true);
-            when(_controller.connectToDevice(any, any, any))
+            when(_controller.connectToDevice(any!, any, any))
                 .thenAnswer((_) => Stream.fromIterable([1]));
 
             _result = _sut.connectToAdvertisingDevice(
@@ -228,15 +228,15 @@ void main() {
 
           test('It emits device update', () {
             expect(_result,
-                emitsInOrder(<ConnectionStateUpdate>[updateForDevice]));
+                emitsInOrder(<ConnectionStateUpdate?>[updateForDevice]));
           });
         });
 
         group('And device is not discovered in a previous scan', () {
           setUp(() {
             when(_scanner.scanForDevices(
-              withServices: anyNamed('withServices'),
-              scanMode: anyNamed('scanMode'),
+              withServices: anyNamed('withServices')!,
+              scanMode: anyNamed('scanMode')!,
             )).thenAnswer((_) => Stream.fromIterable([discoveredDevice]));
           });
 
@@ -276,7 +276,7 @@ void main() {
                       cacheValidity: anyNamed('cacheValidity')))
                   .thenAnswer((_) => responses.removeAt(0));
 
-              when(_controller.connectToDevice(any, any, any))
+              when(_controller.connectToDevice(any!, any, any))
                   .thenAnswer((_) => Stream.fromIterable([1]));
 
               _result = _sut.connectToAdvertisingDevice(
@@ -288,7 +288,7 @@ void main() {
 
             test('It emits device update', () {
               expect(_result,
-                  emitsInOrder(<ConnectionStateUpdate>[updateForDevice]));
+                  emitsInOrder(<ConnectionStateUpdate?>[updateForDevice]));
             });
           });
         });
