@@ -25,14 +25,14 @@ import 'reactive_ble_test.mocks.dart';
 )
 void main() {
   group('$FlutterReactiveBle', () {
-    MockBleOperationController _bleOperationController;
-    MockDeviceScanner _deviceScanner;
-    MockDeviceConnector _deviceConnector;
-    MockConnectedDeviceOperation _deviceOperation;
-    StreamController<BleStatus> _bleStatusController;
+    MockBleOperationController? _bleOperationController;
+    MockDeviceScanner? _deviceScanner;
+    MockDeviceConnector? _deviceConnector;
+    MockConnectedDeviceOperation? _deviceOperation;
+    late StreamController<BleStatus> _bleStatusController;
     MockLogger _debugLogger;
 
-    FlutterReactiveBle _sut;
+    late FlutterReactiveBle _sut;
 
     setUp(() {
       _bleOperationController = MockBleOperationController();
@@ -42,10 +42,10 @@ void main() {
       _bleStatusController = StreamController();
       _debugLogger = MockLogger();
 
-      when(_bleOperationController.initialize()).thenAnswer(
+      when(_bleOperationController!.initialize()).thenAnswer(
         (_) => Future.value(),
       );
-      when(_bleOperationController.bleStatusStream).thenAnswer(
+      when(_bleOperationController!.bleStatusStream).thenAnswer(
           (realInvocation) => _bleStatusController.stream.asBroadcastStream());
 
       _sut = FlutterReactiveBle.witDependencies(
@@ -62,7 +62,7 @@ void main() {
     });
 
     group('BleStatus stream', () {
-      Stream<BleStatus> bleStatusStream;
+      Stream<BleStatus>? bleStatusStream;
       setUp(() {
         bleStatusStream = _sut.statusStream;
       });
@@ -99,10 +99,10 @@ void main() {
       );
       final charValue = CharacteristicValue(
           characteristic: _createChar(), result: characteristic);
-      Stream<CharacteristicValue> charValueStream;
+      Stream<CharacteristicValue>? charValueStream;
 
       setUp(() {
-        when(_deviceOperation.characteristicValueStream)
+        when(_deviceOperation!.characteristicValueStream)
             .thenAnswer((_) => Stream.fromIterable([charValue]));
         charValueStream = _sut.characteristicValueStream;
       });
@@ -119,7 +119,7 @@ void main() {
 
     group('Deinitialize', () {
       setUp(() async {
-        when(_bleOperationController.deinitialize()).thenAnswer((_) async => 1);
+        when(_bleOperationController!.deinitialize()).thenAnswer((_) async => 1);
         await _sut.deinitialize();
       });
 
@@ -130,11 +130,11 @@ void main() {
 
     group('Read characteristic', () {
       QualifiedCharacteristic characteristic;
-      List<int> result;
+      List<int>? result;
 
       setUp(() async {
         characteristic = _createChar();
-        when(_deviceOperation.readCharacteristic(any))
+        when(_deviceOperation!.readCharacteristic(any!))
             .thenAnswer((_) async => [1]);
 
         result = await _sut.readCharacteristic(characteristic);
@@ -152,9 +152,9 @@ void main() {
       setUp(() async {
         characteristic = _createChar();
 
-        when(_deviceOperation.writeCharacteristicWithResponse(
-          any,
-          value: anyNamed('value'),
+        when(_deviceOperation!.writeCharacteristicWithResponse(
+          any!,
+          value: anyNamed('value')!,
         )).thenAnswer((_) async => [0]);
 
         await _sut.writeCharacteristicWithResponse(characteristic,
@@ -173,9 +173,9 @@ void main() {
       setUp(() async {
         characteristic = _createChar();
 
-        when(_deviceOperation.writeCharacteristicWithoutResponse(
-          any,
-          value: anyNamed('value'),
+        when(_deviceOperation!.writeCharacteristicWithoutResponse(
+          any!,
+          value: anyNamed('value')!,
         )).thenAnswer((_) async => [0]);
 
         await _sut.writeCharacteristicWithoutResponse(
@@ -192,10 +192,10 @@ void main() {
     group('Request mtu', () {
       const deviceId = '123';
       const mtu = 120;
-      int result;
+      int? result;
 
       setUp(() async {
-        when(_deviceOperation.requestMtu(any, any))
+        when(_deviceOperation!.requestMtu(any!, any))
             .thenAnswer((_) async => mtu);
 
         result = await _sut.requestMtu(deviceId: deviceId, mtu: mtu);
@@ -211,7 +211,7 @@ void main() {
       const priority = ConnectionPriority.highPerformance;
 
       setUp(() async {
-        when(_deviceOperation.requestConnectionPriority(any, any))
+        when(_deviceOperation!.requestConnectionPriority(any!, any!))
             .thenAnswer((_) async => 0);
         await _sut.requestConnectionPriority(
             deviceId: deviceId, priority: priority);
@@ -234,14 +234,14 @@ void main() {
         rssi: -39,
         serviceData: {},
       );
-      Stream<DiscoveredDevice> deviceStream;
+      Stream<DiscoveredDevice>? deviceStream;
 
       setUp(() {
-        when(_deviceScanner.scanForDevices(
-          withServices: anyNamed('withServices'),
-          scanMode: anyNamed('scanMode'),
+        when(_deviceScanner!.scanForDevices(
+          withServices: anyNamed('withServices')!,
+          scanMode: anyNamed('scanMode')!,
           requireLocationServicesEnabled:
-              anyNamed('requireLocationServicesEnabled'),
+              anyNamed('requireLocationServicesEnabled')!,
         )).thenAnswer((_) => Stream.fromIterable([device]));
 
         deviceStream = _sut.scanForDevices(
@@ -273,11 +273,11 @@ void main() {
       final servicesToDiscover = {
         Uuid.parse('FEFF'): [Uuid.parse('FE1F')]
       };
-      Stream<ConnectionStateUpdate> deviceUpdateStream;
+      Stream<ConnectionStateUpdate>? deviceUpdateStream;
 
       setUp(() {
-        when(_deviceConnector.connect(
-                id: anyNamed('id'),
+        when(_deviceConnector!.connect(
+                id: anyNamed('id')!,
                 servicesWithCharacteristicsToDiscover:
                     anyNamed('servicesWithCharacteristicsToDiscover'),
                 connectionTimeout: anyNamed('connectionTimeout')))
@@ -314,16 +314,16 @@ void main() {
       final servicesToDiscover = {
         Uuid.parse('FEFF'): [Uuid.parse('FE1F')]
       };
-      Stream<ConnectionStateUpdate> deviceUpdateStream;
+      Stream<ConnectionStateUpdate>? deviceUpdateStream;
 
       setUp(() {
-        when(_deviceConnector.connectToAdvertisingDevice(
-          id: anyNamed('id'),
+        when(_deviceConnector!.connectToAdvertisingDevice(
+          id: anyNamed('id')!,
           servicesWithCharacteristicsToDiscover:
               anyNamed('servicesWithCharacteristicsToDiscover'),
           connectionTimeout: anyNamed('connectionTimeout'),
-          prescanDuration: anyNamed('prescanDuration'),
-          withServices: anyNamed('withServices'),
+          prescanDuration: anyNamed('prescanDuration')!,
+          withServices: anyNamed('withServices')!,
         )).thenAnswer((realInvocation) => Stream.fromIterable([update]));
 
         deviceUpdateStream = _sut.connectToAdvertisingDevice(
@@ -353,7 +353,7 @@ void main() {
       );
 
       setUp(() async {
-        when(_bleOperationController.clearGattCache(any))
+        when(_bleOperationController!.clearGattCache(any!))
             .thenAnswer((_) async => result);
 
         await _sut.clearGattCache(deviceId);
@@ -371,10 +371,10 @@ void main() {
         failure: null,
       );
 
-      Stream<ConnectionStateUpdate> updateStream;
+      Stream<ConnectionStateUpdate>? updateStream;
 
       setUp(() {
-        when(_deviceConnector.deviceConnectionStateUpdateStream)
+        when(_deviceConnector!.deviceConnectionStateUpdateStream)
             .thenAnswer((realInvocation) => Stream.fromIterable([update]));
 
         updateStream = _sut.connectedDeviceStream;
@@ -400,11 +400,11 @@ void main() {
       QualifiedCharacteristic char;
 
       Stream<List<int>> valueStream;
-      Stream<List<int>> resultStream;
+      Stream<List<int>>? resultStream;
 
       setUp(() {
         char = _createChar();
-        when(_deviceConnector.deviceConnectionStateUpdateStream)
+        when(_deviceConnector!.deviceConnectionStateUpdateStream)
             .thenAnswer((_) => Stream.fromIterable([update]));
 
         valueStream = Stream.fromIterable([
@@ -412,7 +412,7 @@ void main() {
           [2]
         ]);
 
-        when(_deviceOperation.subscribeToCharacteristic(any, any))
+        when(_deviceOperation!.subscribeToCharacteristic(any!, any!))
             .thenAnswer((_) => valueStream);
 
         resultStream = _sut.subscribeToCharacteristic(char);
@@ -438,7 +438,7 @@ void main() {
         const result = <DiscoveredService>[];
 
         setUp(() {
-          when(_deviceOperation.discoverServices(any))
+          when(_deviceOperation!.discoverServices(any!))
               .thenAnswer((_) async => result);
         });
 
