@@ -3,7 +3,7 @@ import 'package:flutter_test/flutter_test.dart';
 
 void main() {
   group("selectFrom", () {
-    _Enum sut(int raw) =>
+    _Enum sut(int? raw) =>
         selectFrom(_Enum.values, index: raw, fallback: (raw) => _Enum.unknown);
 
     test("selects a value by index", () {
